Stop copying/creating centroids all the time
This patch makes all internal calls handle *centroid instead of the
struct directy. With some crappy profiling in osx, a lot of time was
spent allocating and moving memory around so this seemed like a very
easy improvement to make.

> $ go test -v ./... -bench . -run XXX
> PASS
> BenchmarkUpdate1-4       1000000              1364 ns/op
> BenchmarkUpdate10-4       200000             11540 ns/op
> BenchmarkUpdate100-4       30000             58091 ns/op
> ok      github.com/caio/go-tdigest      6.026s

This is roughly an unscientific 300% improvement in the compression=100
case (which is what should generally be used, according to the original
paper) - Ref: c6047295a555a07c80a203c9b64e20a423cb6fed.