Extract estimateCapacity function
The ideal situation is that we _never_ have to grow the slice,
20 is pretty much a very high upper bound which should only happen
for very low compression numbers or pathological insertions.

I've used a very scienfic approach to come up with the number 10 as
the capacity estimator multiplier (read: I put a panic() when the
slice had to grow and replayed latency numbers from random places.
Started with 5 and went up until it stopped panic()ing then added
one. A.K.A: not scientific at all. ymmv).