Completely rework the quantile estimation codepath
This patch is too big, but there isn't much getting away from it
in smaller steps because summary{} and TDigest{} are actually
tightly coupled (i.e.: the abstraction is mostly useful for
code organization but fails at isolation).

The major changes are:

- Summaries now hold repeated items instead of just unique means
  and their respective counts (which led to changes in how the
  digest adds new centroids too)
- Quantile estimation is now a straight port from the reference
  implementation (issue-84 branch)

The digest now doesn't potentially report completely wrong values
on distributions with multiple steep hills nor biases in favour
of big centroids with few occurrences.

This patch closes #12. Some historical details for the motivation
for this work can be found on PR #11.