Change TDigest.count to uint64 (was uint32)
So now we can register more than 4B samples and still reply with valid
quantile estimations, but notice that centroid counts are still uint32
and that `t.count` is used for floating point divisions across the
algorithms so a really big count is not really desirable.

Individual centroid counts wrapping around uint32 is a pathological
scenario that can totally be simulated and treated for, but that would
mean a lot of extra comparisons added in the critical path, so I prefer
to leave that to the user's decision.