Properly compute extreme CDFs
When we reach the last centroid in the summary, what we want to
do is assume that the last two centroids are of equal width
then estimate the CDF via a simple interpolation. Before this
patch we would wrongly bound (and compute) this estimation with
the last item instead of the one before the last.

Fixes #17